Moved config files to a separate folder
