+++ config/webpack.prod.js
@@ -4,14 +4,14 @@ var webpack = require('webpack');
 var HtmlWebpackPlugin = require('html-webpack-plugin');
 var ExtractTextPlugin = require('extract-text-webpack-plugin');
 
-commonWebpackConfig = require('./webpack.config');
+commonWebpackConfig = require('./webpack.base');
 
 productionWebpackConfig = Object.assign(commonWebpackConfig, {
   cache: true,
   devtool: 'source-map',
   entry: ['babel-polyfill', 'babel-regenerator-runtime', './app/index'],
   output: {
-    path: path.join(__dirname, 'dist'),
+    path: path.join(__dirname, '../dist'),
     filename: 'bundle.[hash].js',
     publicPath: '/static/',
   },
@@ -26,9 +26,7 @@ productionWebpackConfig.plugins.push(
 );
 productionWebpackConfig.plugins.push(
   new webpack.optimize.UglifyJsPlugin({
-    compress: {
-      sourceMap: true,
-    },
+    sourceMap: true,
   })
 );
 productionWebpackConfig.plugins.push(
