+++ package.json
@@ -4,8 +4,8 @@
   "main": "index.js",
   "scripts": {
     "clean": "rimraf dist",
-    "build:webpack": "NODE_ENV=production webpack --config webpack.config.prod.js",
-    "build:analyze": "NODE_ENV=production webpack --config webpack.config.prod.js --json | webpack-bundle-size-analyzer",
+    "build:webpack": "NODE_ENV=production webpack --config config/webpack.prod.js",
+    "build:analyze": "NODE_ENV=production webpack --config config/webpack.prod.js --json | webpack-bundle-size-analyzer",
     "build": "npm run clean && npm run build:webpack",
     "start": "NODE_ENV=production node index.js",
     "dev": "NODE_ENV=development nodemon --watch server index.js",
@@ -49,7 +49,7 @@
     ],
     "setupFiles": [
       "raf/polyfill",
-      "<rootDir>/setupJest.js",
+      "<rootDir>/config/setupJest.js",
       "<rootDir>/__mocks__/window.js"
     ]
   },
